# docs/openapi/generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@livetale.com
    name: LiveTale Team
  description: |
    LiveTale �꾨줈�앺듃�� API 紐낆꽭��
    �명꽣�숉떚釉� �ㅽ넗由ы뀛留곴낵 罹먮┃�� 梨꾪똿�� �꾪븳 醫낇빀 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LiveTale API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.livetale.com/v1
- description: Staging server
  url: https://staging-api.livetale.com/v1
- description: Development server
  url: http://localhost:3000/v1
security:
- BearerAuth: []
- ApiKeyAuth: []
paths:
...
  /user/me:
    get:
      description: �꾩옱 濡쒓렇�명븳 �ъ슜�먯쓽 �뺣낫瑜� 議고쉶�⑸땲��
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: �ъ슜�� �뺣낫 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �� �뺣낫 議고쉶
      tags:
      - User
      x-accepts:
      - application/json
      x-tags:
      - tag: User
...
  /user/achievements:
    get:
      description: �ъ슜�먯쓽 �꾩쟾怨쇱젣 紐⑸줉怨� 吏꾪뻾 �곹솴�� 議고쉶�⑸땲��
      parameters:
      - description: �꾩쟾怨쇱젣 移댄뀒怨좊━ �꾪꽣
        explode: true
        in: query
        name: category
        required: false
        schema:
          enum:
          - story
          - character
          - social
          - exploration
          - creative
          type: string
        style: form
      - description: �꾩쟾怨쇱젣 �곹깭 �꾪꽣
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - in_progress
          - completed
          - locked
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_user_achievements_get_200_response"
          description: �꾩쟾怨쇱젣 紐⑸줉 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �꾩쟾怨쇱젣 由ъ뒪�� & 吏꾪뻾�� 議고쉶
      tags:
      - Achievement
      x-accepts:
      - application/json
      x-tags:
      - tag: Achievement
  /user/achievements/{id}:
    patch:
      description: �뱀젙 �꾩쟾怨쇱젣�� 吏꾪뻾 �곹솴�� �낅뜲�댄듃�⑸땲��
      parameters:
      - description: �꾩쟾怨쇱젣 ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_user_achievements__id__patch_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Achievement"
          description: �꾩쟾怨쇱젣 �낅뜲�댄듃 �깃났
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �섎せ�� �붿껌
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �꾩쟾怨쇱젣瑜� 李얠쓣 �� �놁쓬
      summary: �꾩쟾怨쇱젣 �낅뜲�댄듃
      tags:
      - Achievement
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Achievement
  /user/settings:
    get:
      description: �ъ슜�먯쓽 媛쒖씤 �ㅼ젙�� 議고쉶�⑸땲��
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_user_settings_get_200_response"
          description: �ㅼ젙 議고쉶 �깃났
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �좎� �ㅼ젙 議고쉶
      tags:
      - User
      x-accepts:
      - application/json
      x-tags:
      - tag: User
    patch:
      description: �ъ슜�먯쓽 媛쒖씤 �ㅼ젙�� 蹂�寃쏀빀�덈떎
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_user_settings_patch_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_user_settings_patch_200_response"
          description: �ㅼ젙 �낅뜲�댄듃 �깃났
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �섎せ�� �붿껌
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: �몄쬆�섏� �딆� �ъ슜��
      summary: �좎� �ㅼ젙 蹂�寃�
      tags:
      - User
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: User
...
components:
  schemas:
    User:
      example:
        subscription_tier: free
        is_active: true
        avatar_url: https://openapi-generator.tech
        updated_at: 2000-01-23T04:56:07.000+00:00
        nickname: nickname
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        email: email
        username: username
      properties:
        id:
          description: �ъ슜�� 怨좎쑀 ID
          format: uuid
          type: string
        email:
          description: �대찓�� 二쇱냼
          format: email
          type: string
        username:
          description: �ъ슜�먮챸
          maxLength: 20
          minLength: 3
          type: string
        nickname:
          description: �됰꽕��
          maxLength: 30
          type: string
        avatar_url:
          description: �꾨줈�� �대�吏� URL
          format: uri
          type: string
        created_at:
          description: 怨꾩젙 �앹꽦��
          format: date-time
          type: string
        updated_at:
          description: 留덉�留� �섏젙��
          format: date-time
          type: string
        is_active:
          description: 怨꾩젙 �쒖꽦�� �곹깭
          type: boolean
        subscription_tier:
          description: 援щ룆 �깃툒
          enum:
          - free
          - premium
          - pro
          type: string
      type: object
...
    Achievement:
      example:
        name: name
        description: description
        progress: 8
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        category: category
        unlocked_at: 2000-01-23T04:56:07.000+00:00
        is_completed: true
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        progress:
          maximum: 100
          minimum: 0
          type: integer
        is_completed:
          type: boolean
        unlocked_at:
          format: date-time
          type: string
      type: object
...
    Error:
      example:
        details: "{}"
        error: error
        message: message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      type: object
...
    _user_achievements_get_200_response:
      example:
        total_completed: 0
        achievements:
        - name: name
          description: description
          progress: 8
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          category: category
          unlocked_at: 2000-01-23T04:56:07.000+00:00
          is_completed: true
        - name: name
          description: description
          progress: 8
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          category: category
          unlocked_at: 2000-01-23T04:56:07.000+00:00
          is_completed: true
        total_available: 6
      properties:
        achievements:
          items:
            $ref: "#/components/schemas/Achievement"
          type: array
        total_completed:
          description: �꾨즺�� �꾩쟾怨쇱젣 ��
          type: integer
        total_available:
          description: �꾩껜 �꾩쟾怨쇱젣 ��
          type: integer
      type: object
    _user_achievements__id__patch_request:
      properties:
        progress:
          description: 吏꾪뻾�� (0-100)
          maximum: 100
          minimum: 0
          type: integer
        is_completed:
          description: �꾨즺 �щ�
          type: boolean
      type: object
    _user_settings_get_200_response_notifications:
      example:
        story_updates: true
        email: true
        push: true
      properties:
        email:
          description: �대찓�� �뚮┝ �ㅼ젙
          type: boolean
        push:
          description: �몄떆 �뚮┝ �ㅼ젙
          type: boolean
        story_updates:
          description: �ㅽ넗由� �낅뜲�댄듃 �뚮┝
          type: boolean
      type: object
    _user_settings_get_200_response_privacy:
      example:
        story_sharing: true
        profile_visibility: public
      properties:
        profile_visibility:
          description: �꾨줈�� 怨듦컻 �ㅼ젙
          enum:
          - public
          - friends
          - private
          type: string
        story_sharing:
          description: �ㅽ넗由� 怨듭쑀 �덉슜
          type: boolean
      type: object
    _user_settings_get_200_response_story_preferences:
      example:
        auto_save: true
        image_quality: low
      properties:
        auto_save:
          description: �먮룞 ���� �ㅼ젙
          type: boolean
        image_quality:
          description: �대�吏� �덉쭏 �ㅼ젙
          enum:
          - low
          - medium
          - high
          type: string
      type: object
    _user_settings_get_200_response:
      example:
        privacy:
          story_sharing: true
          profile_visibility: public
        theme: light
        language: ko
        story_preferences:
          auto_save: true
          image_quality: low
        notifications:
          story_updates: true
          email: true
          push: true
      properties:
        theme:
          description: �뚮쭏 �ㅼ젙
          enum:
          - light
          - dark
          - auto
          type: string
        language:
          description: �몄뼱 �ㅼ젙
          enum:
          - ko
          - en
          - ja
          - zh
          type: string
        notifications:
          $ref: "#/components/schemas/_user_settings_get_200_response_notifications"
        privacy:
          $ref: "#/components/schemas/_user_settings_get_200_response_privacy"
        story_preferences:
          $ref: "#/components/schemas/_user_settings_get_200_response_story_preferences"
      type: object
    _user_settings_patch_request_notifications:
      properties:
        email:
          type: boolean
        push:
          type: boolean
        story_updates:
          type: boolean
      type: object
    _user_settings_patch_request_privacy:
      properties:
        profile_visibility:
          enum:
          - public
          - friends
          - private
          type: string
        story_sharing:
          type: boolean
      type: object
    _user_settings_patch_request_story_preferences:
      properties:
        auto_save:
          type: boolean
        image_quality:
          enum:
          - low
          - medium
          - high
          type: string
      type: object
    _user_settings_patch_request:
      properties:
        theme:
          enum:
          - light
          - dark
          - auto
          type: string
        language:
          enum:
          - ko
          - en
          - ja
          - zh
          type: string
        notifications:
          $ref: "#/components/schemas/_user_settings_patch_request_notifications"
        privacy:
          $ref: "#/components/schemas/_user_settings_patch_request_privacy"
        story_preferences:
          $ref: "#/components/schemas/_user_settings_patch_request_story_preferences"
      type: object
    _user_settings_patch_200_response:
      example:
        settings: "{}"
        message: �ㅼ젙�� �깃났�곸쑝濡� �낅뜲�댄듃�섏뿀�듬땲��
      properties:
        message:
          example: �ㅼ젙�� �깃났�곸쑝濡� �낅뜲�댄듃�섏뿀�듬땲��
          type: string
        settings:
          description: �낅뜲�댄듃�� �ㅼ젙 �뺣낫
          type: object
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey